backend/requirements.txt
fastapi
uvicorn
requests
pyshark
openai
python-dotenv